Add concept of a default user.
When using SciTokens with xrootd-hdfs, most filesystem operations will
be performed utilizing the `nobody` user, which may mess up permissions
when accessing HDFS from non-Xrootd-based mechanisms.

With this change, xrootd-scitokens will set a default username that the
filesystem plugin would be able to utilize when interacting with the
storage.  The `default_user` parameter in the issuer section does not
affect mapping decisions.