Improve issuer path logic.
This provides some more flexibility in the path logic for the issuer
configuration:

- Allow `base_path` to be a comma-separated list of paths if the issuer
  operates in different parts of the namespace.
- Add a `restricted_path` option.  This comma-separated-value is a restriction
  on the directory paths the issuer may provide authorizations for inside
  its namespace.

The `restricted_path` option is meant to help sites adopt SciTokens, particularly
those that have setup their filesystem namespace such that the `base_path` is
`/`, but know the issuer will really only write into a single sub-directory (such
as `/store`).  Without this, such a site would have to change their directory layout
(which is disruptive) or allow the issuer to authorize throughout the entire storage.